Implement the stop_after_current API functionality
Triggering this will stop playback after the song that's currently
being played.

Resuming playback (e.g. by calling "play") will start playing the next
song and then just keep going.

This is different from the "single" mode: it only applies to the
currently-playing song. And it will not reset the playback position
back to the beginning of the playlist.

When you turn on the single mode, once the current track ends, the
next track will always be the first one in the playlist (unless the
"repeat" mode is set too).