Fix NoneType exception at the end of playlist
If we ask to stop after the currently-playing track, but that track is
the last one in the playlist, the `next_track` will be set to `None`
and therefore will not have teh `tlid` attribute.

This doesn't actually seems to break any functionality, but it shows a
tracebak unnecessarily and it is an error that should be fixed anyway.